requirements.txt
Flask
gunicorn
requests
datetime
pillow